Move `event_types` code to `_types`
