+++ sentry_sdk/_types.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 try:
     from typing import TYPE_CHECKING
 except ImportError:
@@ -19,13 +21,31 @@ if TYPE_CHECKING:
     from typing import Tuple
     from typing import Type
     from typing import Union
-    from typing_extensions import Literal
+    from typing_extensions import Literal, Required, TypedDict
+
+    class Event(TypedDict, total=False):
+        event_id: Required[str]
+        platform: Required[Literal["python"]]
+        timestamp: Required[Union[str, int, float]]
+        dist: str
+        environment: str
+        errors: list[dict[str, object]]  # TODO: We can expand on this type
+        extra: dict[object, object]
+        fingerprint: list[str]
+        level: Literal["fatal", "error", "warning", "info", "debug"]
+        logger: str
+        modules: dict[str, str]
+        release: str
+        server_name: str
+        tags: Union[
+            list[str], dict[str, object]
+        ]  # Tags must be less than 200 characters each
+        transaction: str
 
     ExcInfo = Tuple[
         Optional[Type[BaseException]], Optional[BaseException], Optional[TracebackType]
     ]
 
-    Event = Dict[str, Any]
     Hint = Dict[str, Any]
 
     Breadcrumb = Dict[str, Any]
